improved logger config
