+++ lib/create.js
@@ -29,21 +29,10 @@ ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 POSSIBILITY OF SUCH DAMAGE.
 </copyright> */
 
+const { logger, setLoggerLevel } = require("./logger");
 const { MinitError } = require("./error");
 const path = require("path");
-const pino = require("pino");
 const fs = require("fs");
-
-// Create logger instance
-const logger = pino({
-    level: "info",
-    transport: {
-        target: "pino-pretty",
-        options: {
-            colorize: true
-        }
-    }
-});
 
 /**
  * Main create function that generates files from templates
@@ -53,6 +42,9 @@ const logger = pino({
  * @returns {Promise} Promise that resolves with the config object
  */
 const create = (module.exports = function (templateName, config, templateModule) {
+    // Set up logger level
+    setLoggerLevel(config.level);
+
     // Store template name in config for later use
     config.templateName = templateName;
