+++ lib/logger.js
@@ -1,0 +1,21 @@
+const pino = require("pino");
+
+// Create logger instance
+// TODO: the log level should be configurable
+const logger = pino({
+    level: "info",
+    transport: {
+        target: "pino-pretty",
+        options: {
+            colorize: true
+        }
+    },
+    sync: true
+});
+
+exports.setLoggerLevel = function (level) {
+    level = ["trace", "debug", "info", "warn", "error", "fatal"].includes(level) ? level : "info";
+    logger.level = level;
+};
+
+exports.logger = logger;
