+++ lib/template-base.js
@@ -35,21 +35,9 @@ const Q = require("q");
 const qfs = require("q-io/fs");
 const os = require("os");
 const process = require('process');
-const pino = require("pino");
+const { logger } = require("./logger");
+const Mustache = require("mustache");
 
-// Create logger instance
-// TODO: the log level should be configurable
-const logger = pino({
-    level: "debug",
-    transport: {
-        target: "pino-pretty",
-        options: {
-            colorize: true
-        }
-    }
-});
-
-var Mustache = require("mustache");
 // We're interpolating inside Javascript and CSS, and so do not want the
 // default HTML escaping
 Mustache.escape = function (string) {
@@ -66,6 +54,7 @@ exports.TemplateBase = Object.create(Object.prototype, {
     addOptions: {
         value: function (command) {
             command.option("-v, --verbose", "Enable verbose logging");
+            command.option("-l, --level [level]", "Log level (default: info)", "info");
             command.option('-p, --package-home [path]',"absolute path to the package's home directory", this.validatePackageHome, this.defaultPackageHome());
             command.option('-d, --destination [path]', this.destinationOptionDescription, Path.join);
             return command;
